+++ src/databricks/sqlalchemy/base.py
@@ -253,6 +253,7 @@ class DatabricksDialect(default.DefaultDialect):
         # TODO: figure out how to return sqlalchemy.interfaces in a way that mypy respects
         return build_pk_dict(pk_name, pk_constraint_string)  # type: ignore
 
+    @reflection.cache
     def get_foreign_keys(
         self, connection, table_name, schema=None, **kw
     ) -> List[ReflectedForeignKeyConstraint]: